Add checkstyle to check Java code against 2103T coding standards
There is no explicit tool to check coding style and keep it under control.

This could lead to mistaken coding style that does not obey the coding standards set by the teaching team.

Adding checkstyle plugin to the project to take advantage of static analysis.

Let's add checkstyle to the project, to prevent the code becoming a spaghetti, with the help of static analysis.