Add GUI on the application using JavaFX
The application support text-based interaction with the user.

This could be difficult for normal users without much affinity to keyboard to use the application, thus decreasing generality of application.

Adding GUI to the application using JavaFX.

The implementation of GUI is not perfect yet, as the implementation only echoes the user input.

Let's add basic GUI setup to the project to lay foundation for further development of GUI on the project.