Use assert feature to document assumptions in the code
The application uses Exceptions to check on expected errors by the user or developer.

However, there may be unexpected errors, especially due to misled or  missed out assumptions in the code, which might introduce unexpected bugs to the program, thus leading the program to collapse.

Adding assertions to document important assumptions that should hold at various points in the code.

The assertion was used to
1. Valide input parameters for private methods
2. Points in the program to ensure validity of a fact which is almost certainly true
3. Validating post condition of the method, before returning the result

Assertions will help the developer detect the bug in development phase, thus preventing the bug to slip-through and cause issue later on in the dev process. It helps finding out problems at earlier stage of development as possible, thus decreasing cost.

Let's add assertions to document assumptions that must hold true within the program.