Configure GitHub Action to build the jar file and automatically generate release
Donji, when a new version is released, or a new feature is added, the changes are not immediately reflected to the latest release. Rather, the developer needs to manually create a new jar file, and attach to a new release.

This can be cumbersome, and does not align with fast delivery to reflect user feedback swiftly, which is crucial to survive in software industry.

Adding Github Action action to automate build and release.

Addition of CI into the project will enable the user of this product to enjoy latest product at anytime, and delivery newest features as soon as it is implemented on the code.

Let's use Github Action to automate build and release.