Change the images for the user and the duke
The image of Dongji and the user was a default image that had less
correlation with what the actual user case context was.

Change the images to be more representative of the user and the Dongji's
nature of the application.

The changed images are a better representation of user and Dongji.