Add help command
Dongji cannot be used without knowing the commands.
However, the commands are not listed anywhere, so
users have to guess the commands or read from the
user guide.

Add a help command that lists all the commands available in Dongji.

Adding help command that lists all the commands will
help users find out how to use the application within
the application itself.

Let's add help command to the application for users
convenience.