Revise README to contain valid contents
The application does not have any documentation for users to find out
how the application is working, and user doesn't know how to interact
with Dongji.

Revise the current README.md file to explain how to use the application.

The README.md will help users to understand how to use the application
more intuitively using human-friendly languages and examples.