Add Testings for various classes in the project
The project does not have extensive testings for the entirity classes,
and does not ensure high coverage for the entire code base.

Add testings for the following classes:
- DeadlineCommand
- EventCommand
- ListCommand
- TodoCommand
- CommandParser

More rigorous testing will help to ensure the correctness of the
project, but requires more testings to be added in the future.